requirements.txt
anyio==3.6.2
APScheduler==3.9.1
asgiref==3.5.2
asn1crypto==1.5.1
Brotli==1.0.9
cachetools==5.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cssselect2==0.7.0
Django==4.1.2
django-cors-headers==3.13.0
fonttools==4.38.0
h11==0.12.0
html5lib==1.1
httpcore==0.14.7
httpx==0.22.0
idna==3.4
oscrypto==1.3.0
Pillow==9.2.0
pycparser==2.21
pydyf==0.5.0
PyMySQL==1.0.2
PyPDF3==1.0.6
pyphen==0.13.0
python-bidi==0.4.2
python-decouple==3.6
python-telegram-bot==20.0a0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tinycss2==1.2.1
tornado==6.1
tqdm==4.64.1
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
weasyprint==57.1
webencodings==0.5.1
zopfli==0.2.1
